Adjusted Test Page and Fixed Bug
Removed some unnecessary variables (which practically had the same purpose as existing ones) and changed how the questions are randomised.
Also fixed a bug where the user's question would be inserted twice, causing the results table to be shown incorrectly. NOTE: No longer an issue because the variable originally used to control this was removed (it deemed unnecessary).